drop speed of flash to 4MHz to ensure it works on all rev. E boards
Turns out that the decoupling capacitor for the flash chip, C7, was
badly laid out, and the return path to GND on the chip is too long.
This results in ringing on the MISO signal, on most (but not all) rev. E
boards.

Full speed (48MHz) and half speed (24MHz) tests result in appreciable
read errors. At 12MHz, no board to date has any errors. Since the
amount of reading and writing is actually quite small, setting the
rate down to 4MHz will have no appreciable effect on the device, and
should remove any worry about it working on all boards.